Fix set indent function

+++ android/src/main/java/com/mahatech/flutter_pax_printer_utility/FlutterPaxPrinterUtilityPlugin.java
@@ -207,7 +207,7 @@ public class FlutterPaxPrinterUtilityPlugin implements FlutterPlugin, MethodCall
       final String status = printerUtility.start();
       result.success(status);
     } else if (call.method.equals("leftIndents")) {
-      short indent = call.argument("indent");
+      int indent = call.argument("indent");
       printerUtility.leftIndents(indent);
       result.success(true);
     } else if (call.method.equals("getDotLine")) {
